testing done

+++ HoopDreams/resolvers/pickupGameResolver.js
@@ -8,7 +8,7 @@ module.exports = {
       new Promise((resolve, reject) => {
         context.PickupGame.find({}, (err, value) => {
           if (err) {
-            reject(customErrors.IntervalServerError());
+            reject(new customErrors.IntervalServerError());
           }
 
           resolve(value);
@@ -18,17 +18,15 @@ module.exports = {
       new Promise((resolve, reject) => {
         context.PickupGame.findById(args.id, (err, value) => {
           if (err) {
-            if ((err.name = "CastError")) {
-              reject(customErrors.NotFoundError);
+            if ((err.name === "CastError")) {
+              reject(new customErrors.BadRequest());
             } else {
-              reject(customErrors.IntervalServerError);
+              reject(new customErrors.IntervalServerError());
             }
           }
+          if(!value) reject(new customErrors.NotFoundError())
           resolve(value);
-        })/*.then(val => {
-          if(val === null) reject(customErrors.NotFoundError);
-
-        });*/
+        })
       })
   },
 
@@ -79,29 +77,42 @@ module.exports = {
 
         }
     },
-    removePickupGame: (root, args, context) =>
-      new Promise((resolve, reject) => {
+    removePickupGame: async (root, args, context) => {
         //first remove from players
         const { id } = args;
 
-        // delete is deprocated used deleteOne
-        context.PickupGame.deleteOne({ _id: id }, (err, removed) => {
-          if (err) {
-            reject(customErrors.NotFoundError);
+        try {
+          const value = await context.PickupGame.findById(id);
+
+          if(!value) {
+            return new customErrors.NotFoundError();
           }
 
-          // then remove player from signupPlayer
-          context.SignupPlayer.deleteMany(
-            { pickupGameId: id },
-            (err, removed) => {
-              if (err) {
-                reject(customErrors.NotFoundError);
-              }
-              resolve(true);
+          // delete is deprocated used deleteOne
+          context.PickupGame.deleteOne({ _id: id }, (err, removed) => {
+            if (err) {
+              return new customErrors.BadRequest();
             }
-          );
-        });
-      })
+            // then remove player from signupPlayer
+            context.SignupPlayer.deleteMany(
+              { pickupGameId: id },
+              (err, removed) => {
+                if (err) {
+                  return new customErrors.NotFoundError();
+                }
+
+              }
+            )
+          })
+          return true;
+        } catch (e) {
+          if(e.name === "CastError") {
+            return new customErrors.BadRequest();
+          } else {
+            return new customErrors.IntervalServerError();
+          }
+        }
+
   },
   addPlayerToPickupGame: (parent, args, context) =>
     new Promise((resolve, reject) => {
@@ -111,42 +122,48 @@ module.exports = {
           reject(err);
         }
         if (new Date() > pickupGame.end) {
-          reject(customErrors.PickupGameAlreadyPassedError)
+          reject(customErrors.PickupGameAlreadyPassedError())
         } else {
           resolve(pickupGame);
         }
 
       });
     }),
-  removePlayerFromPickupGame: (parent, args, context) =>
-    new Promise((resolve, reject) => {
+  removePlayerFromPickupGame: async (parent, args, context) => {
       //context.
       const { input } = args;
-      console.log("input: ", input);
-      context.PickupGame.findById(input.pickupGameId, (err, pickupGame) => {
-        if (err) {
-          reject(err);
-        }
-        if (new Date() > pickupGame.end) {
-          console.log("game has ended");
-          // throw error
-        } else {
-          console.log("date has not ended");
-          context.SignupPlayer.deleteOne(
-            { pickupGameId: input.pickupGameId, playerId: input.playerId },
-            (err, removed) => {
-              if (err) {
-                // throw error
-                reject(err);
+
+      try {
+        const pickupGame = await context.PickupGame.findById(input.pickupGameId);
+          if (!pickupGame) {
+            return customErrors.NotFoundError();
+          }
+          console.log("pickupGame: ", pickupGame);
+          if (new Date() > pickupGame.end) {
+            return new customErrors.PickupGameAlreadyPassedError();
+          } else {
+            const playerSignup = await context.SignupPlayer.find({playerId: input.playerId, pickupGameId: input.pickupGameId});
+            console.log("player: ", playerSignup);
+            
+            if(playerSignup.length === 0) {return new customErrors.NotFoundError();}
+            const del = await context.SignupPlayer.deleteMany(
+              { pickupGameId: input.pickupGameId, playerId: input.playerId },
+              (err, removed) => {
+                if (err) {
+                  // throw error
+                  return err;
+                }
+                return pickupGame;
               }
-              resolve(pickupGame);
-            }
-          );
-        }
-        // check if pickupgame is done
-        resolve(pickupGame);
-      });
-    }),
+            );
+          }
+          return pickupGame;
+
+      } catch (e) {
+        return new customErrors.BadRequest();
+      }
+    }
+  },
   types: {
     PickupGame: {
       registeredPlayers: (parent, args, context) =>
